Desktop/gemini_podcast_kb/requirements.txt
google-genai
feedparser
requests
python-dotenv
streamlit
